==> 1.2_SingleLinearRegression_Deployment.ipynb <==
# Deployment Phase - deploying the model by using pickle class/function
import pickle

loaded_model=pickle.load(open("LinearRegressionModel1_final",'rb')) # load the model by using pickle.load()
result=loaded_model.predict([[0]]) # loading & checking the model with predicted results by inputting values
result # view the output of result
