==> 2.1_MultipleLinearRegression_Model_Creation.ipynb <==
import pandas as pd # Library class/function for using table related datas

dataset=pd.read_csv("2.3_50_Startups.csv") # upload and read csv file data

dataset # view the data table

# for Categorical-values Nominal-data, repeated values in column were distributed as unique values 
# based on one-hot-encoding get_dummies()  

dataset=pd.get_dummies(dataset)

dataset # view data in dataset

# based on one-hot-encoding for categorical values Nominal-data 1st-column need to drop by param drop_first=True

dataset=pd.get_dummies(dataset,drop_first=True)

dataset # View dataset values

dataset.columns # lists column headings

# Split Input data from above data-table and load it to variable as Independent
independent=dataset[['R&D Spend', 'Administration', 'Marketing Spend','State_Florida', 'State_New York']]

independent

# Split Output data from above data-table and load it to variable as dependent
dependent = dataset[["Profit"]]

dependent

# Training_set and Test_Set splitting from input data(independent) & output data(dependent) from SKlearn libraries
# Considering memory consumption calling model_selection function for selecting model 
# Spliting train_set & test_set using train_test_split function 

from sklearn.model_selection import train_test_split

# providing input parameters as independent, output parameters as dependent
# giving training_set data size as 70% & test_set data size as 30% of input and output dataset
train_test_split(independent,dependent,test_size=0.30,random_state=0)

# instead of getting all combined values in a single dataset, based on SKLearn splitting the values as 
# Training_input = x_train, Test_input = x_test
# Training_Output = y_train, Test_output = y_test

x_train,x_test,y_train,y_test = train_test_split(independent,dependent,test_size=0.30,random_state=0)

# view training_set values from the input data Independent
x_train

# view test_set values from the input data Independent
x_test

# view training_set values from the Output data Dependent
y_train

# view test_set values from the Output data Dependent
y_test

# based on SKLearn Library class/function, calling linear_model function & LinearRegression()

from sklearn.linear_model import LinearRegression
regressor=LinearRegression()

# viewing the regressor procedure values
regressor

# substituting train_set values by providing input parameter as x_train & output parameter as y_train by using fit()
regressor.fit(x_train,y_train)

# to view slope/weight by using coef_
weight=regressor.coef_
weight # view the output of weight

# to view bias/intercept by using intercept_
bais=regressor.intercept_
bais # view the output of bais

# Evaluating Metrics by checking model efficiency for test_set values by using predict()
y_pred=regressor.predict(x_test)
y_pred # view the output of y_pred

# Evaluating Metrics values by using r2_score
from sklearn.metrics import r2_score
r2_score # view the output of r2_score

# Evaluating/comparing output values (y_test) and predicted values(y_pred, 
# if result of r_score is near to 1-good Model or near to 0-bad Model
r_score=r2_score(y_test,y_pred)
r_score # view the output of r_score

# saving the model by using pickle class/function

import pickle
filename="MultiLinearRegressionModel_final.sav" # creating dummy filename and saving the model

# dumping the model from regressor by using pickle dump(), open() with write-binary option
pickle.dump(regressor,open(filename,"wb"))

# load and check the Model whether it's working properly by using pickle load(),open() with read-binary option
loaded_model=pickle.load(open("MultiLinearRegressionModel_final.sav",'rb'))
result=loaded_model.predict([[55413,14567,365448,0,1]]) # check and confirm the predicted output result by inputting values
result # view the output of result
