==> 2.2_MultipleLinearRegression_Deployment.ipynb <==
import pickle

loaded_model=pickle.load(open("MultiLinearRegressionModel_final.sav",'rb')) # load the model by using pickle.load()

# loading & checking the model with predicted results by inputting values
result=loaded_model.predict([[55411,145647,355448,0,1]])

result # view the output of result
