==> 3.1_SVM_Model_creation.ipynb <==
import pandas as p # Use pandas library for processing and handling tabularised datas

dataset = p.read_csv("3.3_50_Startups.csv") # upload csv file and read the data using read_csv()
dataset # view the table values

# for Categorical-values Nominal-data, repeated values in column were distributed as unique column-values,
# based on one-hot-encoding using get_dummies()
# here State Column distributed with its unique values and represented as 0/1 (based on one-hot-encoding using get_dummies())
# also 1st-column need to drop by parameter drop_first=True

dataset_1=p.get_dummies(dataset)
dataset_1 # view the table values

# for Categorical-values Nominal-data, repeated values in column were distributed as unique column-values,
# based on one-hot-encoding using get_dummies()
# here State Column distributed with its unique values and represented as 0/1 (based on one-hot-encoding using get_dummies())
# also 1st-column need to drop by parameter drop_first=True

dataset_1=p.get_dummies(dataset,drop_first=True)
dataset_1 # view the table values

dataset_1.columns # list column headings using (variable).columns

# Split Input data from above data-table and load it to variable as Independent
independent=dataset_1[['R&D Spend', 'Administration', 'Marketing Spend','State_Florida', 'State_New York']]
independent # view input values

# Split Output data from above data-table and load it to variable as dependent
dependent=dataset_1[["Profit"]]
dependent# view output values

# Training_set and Test_Set splitting from input data(independent) & output data(dependent) from SKlearn libraries
# Considering memory consumption calling model_selection function for selecting model 
# Spliting train_set & test_set using train_test_split function 

from sklearn.model_selection import train_test_split
x_train,x_test,y_train,y_test=train_test_split(independent,dependent,test_size=0.30,random_state=0)
y_test

from sklearn.preprocessing import StandardScaler
sc=StandardScaler()
x_train=sc.fit_transform(x_train) # fit() calculates & transform-evaluate mean and std deviation value and provides output
x_test=sc.transform(x_test)

# use SKLearn svm and SVR Library class/function
from sklearn.svm import SVR
regressor = SVR(kernel="rbf",C=1000) # check whether kernel is linear or non-linear(rbf-radius bias function)

# substituting train_set values by providing input parameter as x_train & output parameter as y_train by using fit()
regressor=regressor.fit(x_train,y_train)
regressor

regressor.intercept_
regressor.n_support_
regressor.support_
regressor.support_vectors_
#regressor.coef_

y_pred = regressor.predict(x_test)
y_pred

y_test

# use SKLearn metrics - r2_score Library class/function
from sklearn.metrics import r2_score
r_score = r2_score(y_test, y_pred)
r_score
