==> 5.1_RandomForest_Model_creation.ipynb <==
import pandas as p

dataset=p.read_csv("5.3_50_Startups.csv")
dataset

dataset_1=p.get_dummies(dataset,drop_first=True)
dataset_1

dataset_1.columns

independent=dataset_1[['R&D Spend', 'Administration', 'Marketing Spend', 'State_Florida', 'State_New York']]
independent

dependent=dataset_1[['Profit']]
dependent

from sklearn.model_selection import train_test_split
x_train,x_test,y_train,y_test=train_test_split(independent,dependent,test_size=0.30,random_state=0)

from sklearn.ensemble import RandomForestRegressor, RandomForestClassifier
regressor=RandomForestRegressor(n_estimators=50,random_state=0)
regressor.fit(x_train,y_train)

x_test

y_pred=regressor.predict(x_test)
y_pred

from sklearn.metrics import r2_score
r_score = r2_score(y_test,y_pred)
r_score

import pickle
filename="RandomForest_Finalised_Model.sav"
pickle.dump(regressor,open(filename,'wb'))

loaded_model=pickle.load(open("RandomForest_Finalised_Model.sav","rb"))
input
result=loaded_model.predict([[1342,1245,1412,1,0]])
result
