==> 5.2_RandomForest_Deployment.ipynb <==
import pickle

# load RandomForest created Model and open with read access option
loaded_model = pickle.load(open("RandomForest_Finalised_Model.sav",'rb'))

# load the model and compare with predicted inputs  
result = loaded_model.predict([[4356,5678,8741,1,0]])
result
