==> 7.1_GridSearch_DecisionTree_Model_Rscore.ipynb <==
import pandas as p

dataset=p.read_csv("insurance_pre.csv")
dataset

dataset=p.get_dummies(dataset)
dataset

dataset=p.get_dummies(dataset,drop_first=True)
dataset

dataset.columns

# Input dataset

independent = dataset[['age', 'bmi', 'children','sex_female', 'sex_male','smoker_no', 'smoker_yes']]
independent

# Output dataset

dependent = dataset[['charges']]
dependent

# using SKLearn libraries model_selection, spliting input and output values for training-set values & test-set values

from sklearn.model_selection import train_test_split

# providing input parameters as independent, output parameters as dependent
# giving training_set data size as 70% & test_set data size as 30% of input and output dataset

train_test_split(independent,dependent,test_size=0.30,random_state=0)

# instead of getting all combined values in a single dataset, based on SKLearn splitting the values as 
# Training_input = x_train, Test_input = x_test and Training_Output = y_train, Test_output = y_test

x_train,x_test,y_train,y_test=train_test_split(independent,dependent,test_size=0.30,random_state=0)

x_train # to view training - input values

x_test # to view test - input values

y_train # to view training - output values

y_test # to view test - output values

# Applying Support Vector Machine Algorithm 
# based on SKLearn preprocessing Libraries, using StandardScaler()

from sklearn.preprocessing import StandardScaler
sc = StandardScaler()

# fit_tranform() calculates x_train values & transform to evaluate mean and std deviation output values
x_train = sc.fit_transform(x_train)
x_test = sc.transform(x_test)

# based on SKLearn svm libraries, using DecisionTreeRegressor

from sklearn.model_selection import GridSearchCV
from sklearn.tree import DecisionTreeRegressor

# criterion{“squared_error”, “friedman_mse”, “absolute_error”, “poisson”}, default=”squared_error”}
# splitter{“best”, “random”}, default=”best”
# max_features int, float or {“sqrt”, “log2”}

param_grid = {'criterion':['mse','friedman_mse','mae'],
             'splitter':['random','best'],
             'max_features':['auto','sqrt','log2']}

grid = GridSearchCV(DecisionTreeRegressor(), param_grid, refit = True, verbose = 3, n_jobs = -1)

# fitting the model for grid search

grid.fit(x_train,y_train)

# print best parameter after tuning
# print(grid.best_params_)
re = grid.cv_results_

print("The R-Score value for the best parameter {} :".format(grid.best_params_))


# View in Table form

table = p.DataFrame.from_dict(re)
table

# Check and verify

age_input=float(input("Age :"))
bmi_input=float(input("BMI :"))
children_input=float(input("Children :"))
sex_female_input=int(input("Sex: If female give value as 1 otherwise 0 -"))
sex_male_input=int(input("Sex: If male give value as 1 otherwise 0 -"))
smoker_no_input=int(input("Smoker: If non-smoker give value as 1 otherwise 0 -"))
smoker_yes_input=int(input("Smoker:If smoker give value as 1 otherwise 0 -"))

Future_Prediction = grid.predict([[age_input,bmi_input,children_input,sex_female_input,sex_male_input,smoker_no_input,smoker_yes_input]])
# Future_Prediction
print("Future_Prediction={}".format(Future_Prediction))

# # saving the model by using pickle class/function

# import pickle
# filename="SupportVectorMachine_Model.sav"

# # dumping the model from regressor by using pickle dump(), open() with write-binary option
# pickle.dump(regressor,open(filename,'wb'))

# # load and check the Model whether it's working properly by using pickle functions-load(),open() with read-binary option
# loaded_model = pickle.load(open('SupportVectorMachine_Model.sav','rb'))

# # check and confirm the predicted output by inputting multiple input-values
# result = loaded_model.predict([[25,27,0,1,0,1,0]])
# result
